Updated mohit experiments to serve multiple models

+++ llms/scripts/run_mohit_experiments_slurm.sh
@@ -197,26 +197,30 @@ if [ -z "$remote_model"  ]; then
                                     --intervention-utility "$utility" --teacher-explanation-type "$teacher_expl" --student-explanation-type "$student_expl" --use-explanations \
                                     --use-gold-label --budgets "${budgets[@]}" --llm-lib "$lib" --temperature "$gen_temperature" --n-logprobs "$num_logprobs" > "$out_file"
 else
-  echo "Serving student model using $lib"
-  echo "Student's gpus: $student_gpus"
-  CUDA_VISIBLE_DEVICES="$student_gpus" vllm serve "$student_model" --download-dir "$cache_dir" --dtype auto --api-key "$api_key" --gpu-memory-utilization "$gpu_usage" \
-                                                  --tensor-parallel-size "$n_student_gpus" --host "$student_host" --port "$student_port" &
-  student_id=$!
-  sleep 5m
-  echo "Serving teacher model using $lib"
-  echo "Teacher's gpus: $teacher_gpus"
-  CUDA_VISIBLE_DEVICES="$teacher_gpus" vllm serve "$teacher_model" --download-dir "$cache_dir" --dtype auto --api-key "$api_key" --gpu-memory-utilization "$gpu_usage" \
-                                                  --tensor-parallel-size "$n_teacher_gpus" --host "$teacher_host" --port "$teacher_port" &
-  teacher_id=$!
-  sleep 5m
-  printf "Launching Mohit\'s experiment script"
+  if [ "$lib" = "vllm" ]; then
+    echo "Serving student model using vLLM"
+    echo "Student's gpus: $student_gpus"
+    CUDA_VISIBLE_DEVICES="$student_gpus" vllm serve "$student_model" --download-dir "$cache_dir" --dtype auto --api-key "$api_key" --gpu-memory-utilization "$gpu_usage" \
+                                                    --tensor-parallel-size "$n_student_gpus" --host "$student_host" --port "$student_port" &
+    student_id=$!
+    # sleep 5m
+    echo "Serving teacher model using vLLM"
+    echo "Teacher's gpus: $teacher_gpus"
+    CUDA_VISIBLE_DEVICES="$teacher_gpus" vllm serve "$teacher_model" --download-dir "$cache_dir" --dtype auto --api-key "$api_key" --gpu-memory-utilization "$gpu_usage" \
+                                                    --tensor-parallel-size "$n_teacher_gpus" --host "$teacher_host" --port "$teacher_port" &
+    teacher_id=$!
+    sleep 5m
+  fi
+  echo "Launching Mohit's experiment script"
   python src/mohit_mm_experiments.py --data-dir "$data_dir"/"$dataset_dir" --cache-dir "$cache_dir" --train-filename "$train_file" --test-filename "$test_file" \
                                     --val-filename "$val_file" --results-path "$results_path" --task "$dataset" --student-model "$student_model" \
                                     --teacher-model "$teacher_model" --max-new-tokens 100 --n-beams 4 --n-ic-samples 5 --mm-type "$mental_model" \
                                     --intervention-utility "$utility" --teacher-explanation-type "$teacher_expl" --student-explanation-type "$student_expl" --use-explanations \
-                                    --use-gold-label --budgets "${budgets[@]}" --llm-lib "$lib" --remote --student-model-url "$student_model_url" --teacher-model-url "$teacher_model_url" --api-key "$api_key" --temperature "$gen_temperature" \
-                                    --n-logprobs "$num_logprobs" > "$out_file"
-  kill -9 "$student_id" "$teacher_id"
+                                    --use-gold-label --budgets "${budgets[@]}" --llm-lib "$lib" --remote --student-model-url "$student_model_url" --teacher-model-url "$teacher_model_url" --api-key "$api_key" \
+                                    --temperature "$gen_temperature" --n-logprobs "$num_logprobs" > "$out_file"
+  if [ "$lib" = "vllm" ]; then
+    kill -9 "$student_id" "$teacher_id"
+  fi
 fi
 
 conda deactivate
